chore(HandCard) : (int) 명시 추가

+++ src/main/java/domain/HandCard.java
@@ -16,7 +16,7 @@ public class HandCard {
     
     public int cardCalculator(){
         int sum = cards.stream().map(Card::getRankScore).filter(e -> e != 1).mapToInt(Integer::intValue).sum();
-        int ace_cnt = cards.stream().map(Card::getRankScore).filter(e -> e == 1).count();
+        int ace_cnt = (int) cards.stream().map(Card::getRankScore).filter(e -> e == 1).count();
         int result = sum + aceCalculator(sum, ace_cnt);
 
         if(result > BLACKJACK_MAX_LIMIT) return 0;
@@ -27,7 +27,7 @@ public class HandCard {
     private int aceCalculator(int sum, int ace_cnt){
         int totalSum = sum + (ace_cnt*ACE_MAX_VALUE);
         while(totalSum > BLACKJACK_MAX_LIMIT && ace_cnt > 0){
-           totalSum -= (ACE_MAX_VALUE-ACE_MIN_VALUE);
+           totalSum -= (ACE_MAX_VALUE - ACE_MIN_VALUE);
            ace_cnt--;
         }
         return totalSum - sum;
